fix: move Docs code from _.vue
